CodeGen from PR 34394 in Azure/azure-rest-api-specs Merge 4df6f2254ab96439aeff4d425d44708a5b11d003 into b7881d9ab6bb96d1b0c028b1c15b339e5dc010bd
