+++ sdk/durabletask/ci.mgmt.yml
@@ -1,7 +1,6 @@
 # NOTE: Please refer to https://aka.ms/azsdk/engsys/ci-yaml before editing this file.
 
 trigger: none
-
 pr:
   branches:
     include:
