+++ sdk/eventgrid/Azure.Messaging.EventGridV2/CHANGELOG.md
@@ -1,0 +1,11 @@
+# Release History
+
+## 1.0.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
