+++ sdk/eventgrid/Azure.Messaging.EventGridV2/tsp-location.yaml
@@ -1,4 +1,4 @@
-directory: specification/informatica/Informatica.DataManagement.Management
+directory: specification/eventgrid/Azure.Messaging.EventGrid
 commit: 57f84f7b5bcf4b4b53adf2822fe7497615e3d872
 repo: Azure/azure-rest-api-specs
 additionalDirectories:
