--- sdk/pineconevectordb/ci.yml
@@ -1,35 +1,0 @@
-# NOTE: Please refer to https://aka.ms/azsdk/engsys/ci-yaml before editing this file.
-
-trigger:
-  branches:
-    include:
-    - main
-    - hotfix/*
-    - release/*
-  paths:
-    include:
-    - sdk/pineconevectordb
-    - sdk/pineconevectordb/ci.yml
-    - sdk/pineconevectordb/Azure.ResourceManager.PineconeVectorDb
-
-pr:
-  branches:
-    include:
-    - main
-    - feature/*
-    - hotfix/*
-    - release/*
-  paths:
-    include:
-    - sdk/pineconevectordb
-    - sdk/pineconevectordb/ci.yml
-    - sdk/pineconevectordb/Azure.ResourceManager.PineconeVectorDb
-
-extends:
-  template: /eng/pipelines/templates/stages/archetype-sdk-client.yml
-  parameters:
-    ServiceDirectory: pineconevectordb
-    ArtifactName: packages
-    Artifacts:
-    - name: Azure.ResourceManager.PineconeVectorDb
-      safeName: AzureResourceManagerPineconeVectorDb
